*BUG FIX: Viewing a task list was causing an exception sometimes.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3181 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/User Controls/UserControlTasks.cs
@@ -1062,7 +1062,7 @@ namespace OpenDental {
 
 		private void tree_MouseDown(object sender,System.Windows.Forms.MouseEventArgs e) {
 			for(int i=TreeHistory.Count-1;i>0;i--) {
-				if(TreeHistory[i].TaskListNum==(int)tree.GetNodeAt(e.X,e.Y).Tag) {
+				if(TreeHistory[i].TaskListNum==(long)tree.GetNodeAt(e.X,e.Y).Tag) {
 					break;//don't remove the node click on or any higher node
 				}
 				TreeHistory.RemoveAt(i);
